Extract pod allocation manager from status manager
