+++ pkg/kubelet/allocation/state/state.go
@@ -18,6 +18,7 @@ package state
 
 import (
 	v1 "k8s.io/api/core/v1"
+	"k8s.io/apimachinery/pkg/types"
 )
 
 // PodResourceAllocation type is used in tracking resources allocated to pod's containers
@@ -48,6 +49,8 @@ type writer interface {
 	SetContainerResourceAllocation(podUID string, containerName string, alloc v1.ResourceRequirements) error
 	SetPodResourceAllocation(podUID string, alloc map[string]v1.ResourceRequirements) error
 	Delete(podUID string, containerName string) error
+	// RemoveOrphanedPods removes the stored state for any pods not included in the list of remaining pod UIDs.
+	RemoveOrphanedPods(remainingPods map[types.UID]bool)
 }
 
 // State interface provides methods for tracking and setting pod resource allocation
